feat(110642): cria arquivo geral de parsers e move funções
